requirements.txt
asyncoro==3.5
chartkick==0.4.2
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-loklak-api==1.0.1
requests==2.8.1
Werkzeug==0.11.5
wheel==0.26.0
